Service file moved to different locations
